requirements.txt
streamlit
clarifai
python-dotenv
Pillow